Move SyncRecord to models
